OF-2858: Fix error condition for MUC PM
When sending a private message to a user that's not an occupant of the room, XEP-0045 section 7.5 defines that a different error condition is to be used than the one that Openfire has been using. This commit corrects the error condition.

(cherry picked from commit 3b823f851987255b5d85e0c55b1531b8b82b81bf)