OF-2860: Rename MUCRole variables
In preparation of the renaming of the MUCRole class, this commit renames all variables (and some comments / log messages). This aims to explicitly distinguish between the XEP-defined term 'role' and the usage of MUCRole, which is now mostly referenced as 'occupant'.

This commit is not supposed to introduce any functional change.