OF-2862: Refactor MUC 'role' and 'affiliation' usage
Many privileges in multi-user chat are tied to either ‘roles’ or ‘permissions’, as defined in XEP-0045. Openfire’s existing implementation bases the ‘MUCRole' implementation as the source of much. As the ‘MUCRole' class is available only when a user is currently in the chat room, this prevents functionality that’s aught to be allowable when a user is not.

Refactor Openfire’s implementation to explicitly base permission handing based on roles and affiliations.