Remove code marked for deprecation in 4.9.0 or later
Code that was marked for removal in version 4.9.0 should be removed.

Note that we shouldn’t remove code when the mark was applied in versions 4.8.1 or later. In those cases, lets bump up the removal to the next minor version (4.10.0). This allows developers to have one ‘full’ release cycle to be aware of deprecation, and adjust their code accordingly.