OF-2872: Fix XML parsing when previous buffer couldn't be fully parsed
The previous code failed to properly update the read index of the ByteBuffer that is being read, in cases where the ByteBuffer that is parsed had an initial read index higher than 0.

Through manual testing, I've observed this to fix an issue where client connections that send Cyrillic characters get disconnected with a "Stopped parsing never ending stanza" exception.