OF-2785: Happy Eyeballs: Have concurrent outbound connection attempts
When establishing an outbound S2S connection, Happy Eyeballs defines that concurrent connection attempts should be made, when the target domain advertises multiple server sockets (throught DNS SRV etc).

This change implements such concurrency. The first connection that is established will be returned, cancelling other scheduled attempts and aborting attempts that are inflight.

Note that this commit does _not_ address the issue where, when establishing an outbound connection, resolving the address to be used from the XMPP domain name results in a socket connection, that's then immediately closed only to be replaced by a new connection to the same address (OF-2721).