OF-2784: Admin Console's DNS check improvements
Now checks for the presence of A (IPv4) and AAAA (IPv6) records, allowing users to assert if their DNS configuration is correct.

The index page (which pops up a warning to look at the analytics-page if needed) now operates based on a timeout. This prevents slow DNS resolution from keeping the index-page to load in a timely manner. This functionality depends on the new feature introduced by OF-2885.