OF-2784: Admin Console Registration & Login settings to include IPv6
This reworks the login restrictions that can be configured on the Registration & Login page. The text area that was used to fill out _all_ IP addresses / ranges is now replaced with a table and form that takes individual IP addresses and ranges. Support for IPv6 addresses (and ranges) has been added.

All pre-existing notation forms remain functional. The wildcard-based annotation form to denote an IPv4-range (eg `198.41.100.*`) now gets translated in a CIDR variant (`198.51.100.0/24`) when saved. The wildcard notation is _not_ supported for IPv6.