OF-2784: Admin Console Access settings to include IPv6
This reworks the restrictions that can be configured on the Admin Console Access page. The text area that was used to fill out _all_ IP addresses / ranges is now replaced with a table and form that takes individual IP addresses and ranges. Support for IPv6 addresses (and ranges) has been added.