OF-2890: Fix CSRF error on admin setup (directory services)
This fixes a bug that prevented the successful password-check of a newly configured admin.

The page that performs the test is embedded as a dialog in another page. To prevent collisions with the 'csrf' parameter, the parameter on the embedded page is renamed to 'embeddedcsrf'. CSRF processing on the parent page is not performed if the page reload related to the embedded page (CSRF checking is performed there).