OF-2896: Fix memory leak when dealing with pre-authenticated Sessions
A collection of pre-authenticated sessions was maintained incorrectly (using a type-unsafe key that was both used as a stream-ID and a JID), which caused every session that was ever created to remain a part of that collection, until the server got restarted. This introduces a memory leak (and possibly functional issues, although those have never been reported).

This commit resolves the problem. It hides access to the collection to enforce proper maintenance. A more type-safe key is used, which should help prevent future issues like these.

The changes in the second commit affect the API. I've opted to not include backwards-compatibility (with deprecation), as I believe the per-existing API to be simply broken. Any usage of that API should be revisited, which is enforced by removing the API.