OF-2897: Consider anonymous users when evaluating no-such-user
A user that is authenticated using the 'ANONYMOUS' SASL mechanism isn't a _registered_ user, but is a valid user of the system.

The bug fixed in this commit was likely hidden from view by the bug fixed in OF-2896.

(cherry picked from commit cd2ef4f182df6d4bbac07a3d2050b3aef719dbcb)