[Jetty-upgrade] Apply Joakim's suggestions to TempFileToucherTask
As part of the upgrade from Jetty 10 to 12.0.14, the task that prevents Jetty files stored in a temporary directory from being removed needed an upgrade.

Joakim Erdfelt, a Jetty developer, suggested the change in this commit (but also correctly noted that these files ideally aren't stored in a temporary directory, but in a kind of work directory. That would be a larger change. For now, lets at least get the existing mechanism working).