OF-2913: Fix db statistics
The end-time of a profiling session is now observed when calculating rates. Prior to this change, no new data was recorded, but the end-time that was used was 'now', leading to decreasing rates over time.

When (re)starting a profiling session, all old data should be discarded, to avoid skewed averages and rates.