feat: Adds option to tombstone a MUC room on deletion
Tomb-stoning a room means that this room's name is retired on room deletion and cannot be used again. Where a MUC may have particular meaning to plugins or to other systems, assume that deletion is explicit and intentional. Do not allow the MUC name (and so JID) to be repurposed, preventing impersonation / corruption / access of MUC history, or other unforeseen side effects.