OF-2918: Add option to clear history for a given MUC
An improvement to SequenceManager has been made to better support its usage in plugins like the Monitoring plugin.

To use SequenceManager to keep track of and generate new unique IDs, an explicit type number is required. One way to achieve this is to assign a JiveID to a class that uses SequenceManager, e.g. @JiveID(604). However, there is potential for undetected ID clashes if IDs are explicitly assigned in more than one plugin.
This commit proposes a solution that uses unique names to assign a unique SequenceManager type. To obtain a SequenceManager for the roomIDs in the Monitoring plugin, you simply need to call: SequenceManager.getSequenceManagerByUniqueName("Monitoring Plugin - RoomID").

These changes support fixes for issues #369 and #370 in the Monitoring plugin.