OF-2925: Have 'multi' providers for Groups
As are available for Users, UserProperties and Auth, this commit now introduces a Hybrid and Mapped provider for Groups.

With these providers, groups can be obtained from more than one external system.

This change is a prerequisite for OF-2923. As both issues were developed at the same time, some concepts related to both issues are applied to this commit. This foreshadows more, similar changes related to OF-2923.