OF-2924: Unit tests for 'multi' providers
This adds unit tests for most of the 'Hybrid' and 'Mapped' providers for User, UserProperty, Auth and Group.

One bigger issue introduced by the recent (as of yet unreleased) refactoring was identified through these tests, and was fixed in `HybridUserPropertyProvider`

Various smaller issues (mostly with throwing one type of exception while another was expected) have also been addressed by this.