OF-2918: Expose numeric MUC room ID in events
A MUC room of the same name (JID) can be created, destroyed and recreated again. The two rooms should be seen as two different rooms.

When using the JID of the rooms, there's no way to distinghuish between them. For that, Openfire's numeric ID can be used.

In this commit, the numeric ID is added to selected MUC events.