Speed up LocalIncomingServerSessionTest
The test teardown waits for inbound stream IDs to be registered with the session manager. Many streams (those that have errored out or have been replaced with an encrypted stream) won't ever be registered in the Session Manager. The test teardown doesn't need to wait for those.

This commit introduces a change that tracks the streams that are still in-flight. Test teardown will now only wait for those.