OF-2921: Async session close event handlers
This revisits the fix for OF-2921, which introduces a problem in which a broadcast of presence occurs while the session is removed from the session manager. This causes issues, as the broadcast needs data associated to the session (such as its roster, privacy lists, etc).

To fix this, the asynchronous nature of the solution introduced in OF-2921 is revisted. The event handlers that work on session-close events are made asynchronous, to reduce the likelyhood of thread deadlocks.

This commit introduces a new method in the ConnectionCloseListener interface, that intends to replace the pre-existing method. The new method has a default implementation that invokes the old method. This should make for a mostly backwards compatible upgrade.