OF-2950: Remote Session Locator to ignore requests for local sessions
A bug in Openfire has caused a state inconsistency, that leads the local node to not find a session in its own local data. Openfire therefor invokes the ‘session locator’, to find it on other cluster nodes. It supplies the node-ID of the cluster node that is the owner of the session, which in this case, is itself.

The ‘remote’ representation that is unconditionally returned by this locator doesn’t work (as the session isn’t, in fact, remote), leading to confusing log statements that appear to refer to remote sessions, while in fact, the problem is local. These messages have been observed on a cluster of one node.

The session locator routines should be modified to handle these cases more gracefully. They should log a warning, and not return the ‘remote’ surrogate.