Revert "OF-2921: Prevent deadlock by not broadcasting synchronously"
This reverts commit b8b5713c550b3b114020d0b5eb0db4c50660683e.

Routing the unavailable presence that's sent out when a session closes cannot be made asynchronous. If the 'next step' is done prior to the routing of unavailable presence having completed, the session seems to be (partially) re-added. This causes an invalid state: the routing table no longer has the session, but caches such as the Routing Users Cache does.

This incorrect behavior was introduced by a fix for OF-2921.