OF-2793: Bundle an Oracle database driver
Given changes in the Oracle license, we can now bundle their database drivers with Openfire. The changes in this commit do just that.

I've opted to use ojdbc11-production rather than ojdbc17-production as the latter is documented to be compatible with Jakarta. We do not use Jakarta. ojdbc11 has been documented to be compatible with Java 17 and 21, which should be good enough for us for now.

Jetty's JSP compiler plugin is not happy with the standard Oracle dependencies. It fails on expecting but not finding oracle.i18n.util.LocaleMapper. The internet suggests that this is caused by Oracle's xmlparserv2 dependency. Excluding that 'fixes' the problem. I'm not sure what functionality this affects.

As a side-effect, the `lib` folder of a distribution now contains an unneeded POM file. I can't find an easy way to exclude that.