OF-2722: Remove unused plugin download code
By moving JSP code into a servlet, static analyzers are finding more issues. One of the reported issues based on the change in the previous commit is that unsanitized user-provided input is used in the code that implements the downloading of plugins.

This particular code isn't used at all by the admin console. I've removed it in this commit.