OF-2956: Fix database table ofSASLAuthorized upgrade scripts
We're removing this database script in upgrade 37, but we do that after reports of the original definition causing problems: on Oracle, under certain configuration, the primary key index on a 4000-wide column won't work.

In the off-chance that people upgrading in the future will still execute (very) old upgrade scripts (10 and 19), those scripts are modified by this commit to no longer introduce the problem. As 37 will delete the unused table anyway, those changes should be inconsequential.