OF-2960: Replace Oracle database driver's BOM project with individual dependencies
This works around an issue that creeps up when using the BOM project: it causes a POM file to be included in Openfire's distirbution. That file in turn causes the RPM build to fail.

As the distribution cannot be easily changed, this commit prevents usage of the project that introduces the POM file.