OF-2966: BOSH 'close' executed by worker threads
By having the 'close' routine be executed by the same worker threads that process regular BOSH data, the processing order of all of these will make use of the FIFO mechanism introduced in the last two commits.

This helps ensure that a 'close' (terminate) event is processed only after stanzas that were received earlier have been processed - assuming that this processing is synchronous.