OF-3026: Splitting up LocalRoutingTable
Instead of having one 'local' routing table to track routes from all types of connections (component, server, client) a distinct table for each of these types should be used.

This allows for more granular control of data access. It will become easier to move invocations under a guard of a lock form a corresponding cache.

In this commit, only minor functional changes are applied. This affects some exception handling (a more generic `Throwable` is caught when processing of a stanza fails) and the registration of a task that cleans up server sessions (this task is now cancelled when the RoutingTable is stopped).

Changes are applied to classes and methods that all have limited visibility outside of the RoutingTableImpl class (or its package). It's therefor unlikely that this change introduces compilation issues in other projects (such as Openfire plugins). The code changed here isn't expected to be re-used directly by other projects.

(cherry picked from commit 701d2bb797cd2bc43353105632e29c14a48b1759)