OF-3026: State changes of LocalRoutingTable should happen under guard of a lock
When modifications to a LocalRoutingTable is made, this should always happen under guard of a lock that also guards related changes to the corresponding cache.

This commit should reduce the possibility of state inconsistencies. It does so by re-using the Cache-sourced locks that were originally introduced by the changes for issue ENT-425.

In the past, we introduced deadlock potential by putting more operations under guard of a lock. From history, I don't think we ever attempted this change. Additionally, as the code that's now moved under guard of the pre-existing lock is not (intended to be) used by code outside of this package (the class has a corresponding access modified), this change should have minimal risks of introducing such a deadlock.

(cherry picked from commit 0b4c855846811cd75d0c08e01d82adb9feb34e19)