OF-3026: LocalClientRoutingTable addRoute and removeRoute consistency
The argument used to add or remove a route to/from localClientRoutingTable should be created in the same way, for consistency.

Note that this is a non-functional change, as in practice, the `toString` and `toFullJID` invocations would have resulted in the same value. An earlier check ensures that the only difference in those two methods (checking if the resource part is not null) isn't a factor.

(cherry picked from commit 5e8c012379d7306ec805cb231b05228b36b017b3)