chore: replace empty string comparision with `isEmpty()`
Various bits of automation keep insisting that using `isEmpty()` is generally more readable and efficient than comparing against an empty string (`""`). I don't particularly feel that this is a worthwhile change in itself, were it not for it being a good way to reduce the number of 'hints' that are generated from our code base. That in itself may help to reduce noise: it can make other (more important) hints stand out better.

This should not introduce any functional change.