OF-3027: Give Netty threads a recognizable name
To facilitate debugging, the various threads used by Netty EventLoops should have names that at least contain the type of connection that they're operating for.

This commit changes the name of threads of the
- 'parent' eventloop from `nioEventLoopGroup-7-1` (a Netty default) to `socket_s2s_ssl-acceptor-7`
- 'child' eventloop from `socket_s2s_ssl-thread-3` (a Netty default) to `socket_s2s_ssl-worker-3`

To do this, a new theadpool factory is introduces for the 'parent' eventloop. The new factory is configured to use the same priorty and daemon settings as the default Netty threadpool factory.