OF-3029: Prevent false warnings about data inconsistency
It appears that this warning is also logged (incorrectly) in a scenario where user has a session that relates to a different route (in other words: when the user is using more than one device / the addressing uses different resource-parts in their JIDs). Openfire should not be logging that warning in such cases.