OF-3031: Session's 'close' state no longer based on its backing connection
In the early days, a Session was always backed by a (socket) connection. When the connection was closed, so was the session, and vice versa.

With Stream Management, this no longer is the case. A session can remain unclosed (detached) while its connection is closed.

This commit changes the behavior of `LocalSession#isClosed()` to be based on the session's `status` field, and no longer on the connection's connectivity state.