OF-3031: LocalSession isClosed 'old' implementation through configuration
The change made to `LocalSession#isClosed()` under OF-3031 seems to be the 'correct' change, but this change is likely to introduce issues with code that depends on the old behavior.

To allow for backwards compatibility, this commit introduces a configuration property `xmpp.session.isclose.connectionbased` that can be used to restore the old behavior. By default, the old behavior will be disabled, but this can be changed by changing the value of this property to `true`.