OF-3035: Propagate session 'detached' state through cluster
This commit ensures that all cluster nodes can see the 'detached' state of a session. Prior to this change, this was only visible at the local cluster node. Because of this, cluster tasks that cannot operate on detached sessions no longer are invoked.

Additionally, certain cluster tasks that cannot succeed on a detached session are now logged in a less verbose manner than on level 'error'.