OF-3039: Avoid resource conflict resolution on closed session
When a session is closed, there is no need to apply resource-part conflict resolution anymore (as that session isn't going to be used anyway).

This can also prevent a cluster task exception: the cluster task to increase a conflict count will throw an exception when the session being operated on is non-existing, or existing-but-closed.