OF-3038: Reduce verbosity of Socket Connection reset log messages (c2s only)
Clients, and IOS-based clients in particular, can be expected to drop their TCP connection a lot. This event currently logs a stack trace (shown below). For an event that is very expected, this is to verbose. Lets reduce the verbosity of this event.

I’d prefer to keep the verbosity for other connection types than client-to-server connections. Remote servers, for example, are not expected to close the TCP connection ‘uncleanly'.