OF-3041: Prevent Deadlock by reworking `LocalSession#canDeliver()`
Introduces an alternative to `LocalSession#canProcess`: `#canDeliver`, to fix:
- no longer generate two error stanzas if `LocalOutgoingServerSession#canProcess` returns `false`
- have implemenations be responsible for generating any error stanzas, making it possible for each implemenation to choose the best type of error. No longer assume that every error must be related to XEP-0016.
- prevent synchronous error processing in `LocalOutgoingServerSession` that can introduce a deadlock (similar to OF-2341 and OF-2342).

This commit replaces `#canProcess` but marks it as being deprecated. The replacement method, `#canDeliver` has temporarily received a default implementation that is backwards compatible with the preexisting code. This should allow any third-party implementations to be able to gracefully migrate. In future versions of Openfire `#canProcess` will be removed, and the definition of `#canDeliver` will be changed: instead of a default implementation, it will only be an interface definition of a method (`public boolean canDeliver(@Nonnull final Packet stanza);`)

(see https://igniterealtime.atlassian.net/browse/OF-3041 for more details)