OF-2607: New database schema for XEP-0421
This adds a database update for MUC services to contain a private key value. This value is intended to be used by an implementation of XEP-421 'Occupant identifiers for semi-anonymous MUCs' that will be added in commits following this one.

One change is added in the database change (for which both the 'install' script as an 'update' script has been provided):
- a new column is added to the `ofMucService` table
- for each pre-existing row of that table, a distinct random text value is added.