OF-2607: Add 'privateKey' property to MUC Services
In preparation of the addition of support for XEP-0421 'Occupant identifiers for semi-anonymous MUCs', each MultiUserChatService now has a privateKey field.

The changes in this commit ensure that this new property is properly stored in/obtained from database storage. New services get a randomly created value. As Occupant-id values are expected to be stable, privateKey values should not change. Modification of that value is thus not supported by the changes applied in this commit.