CI: Introduce checks for the database install scripts
Prior to this commit, Openfire's continuous integration already had some scripts that verify the scripts used to update the database managed by Openfire from some base version to a newer version. These checks do not verify the installation script - only the update scripts, which provide a 'delta'.

In this commit, the mechanism used to test the update scripts is repurposed to also test the install scripts:
- instead of provisioning the database with a known-good (but old) database installation script, the latest version of the script is used (which is the system under test)
- the 'update runner' is not executed at all - instead, it is assumed that a bug in the install script will cause the docker build to fail.