OF-3044: ReverseLookupUpdatingCacheEntryListener optimization for unique ownership
Some instances of ReverseLookupUpdatingCacheEntryListener are used in contexts where it's expected that exactly one cluster node 'owns' the data for which an event was fired.

For such scenarios, this commit adds an optimization that removes data registered for _other_ nodes, when data is being registered for a node. This intends to add some resilience against data inconsistency.