OF-3048: Improve ClientSession comparison performance (clustering)
Comparing ClientSessions typically involves evaluating the 'is anonymous' property of a session.

In a cluster, the property value of remote sessions is retrieved under guard of a distributed lock. As comparison of larger sets frequently evaluates the property value, that lock is obtained very frequently. This dramatically reduces performance.

As the property will never change after authentication, it's safe to cache the value. This improves performance considerably.

Note that RemoteClientSession instances are typically short-lived (as they're typically discarded after use), which reduces the effectiveness of this change. Still, even without re-use, the performance improvement is considerable: in a cluster with 10,000 sessions, the responsiveness of the session summary page is improved from many minutes to 20 to 30 seconds in my test environment.

(cherry picked from commit 14832dacf84faf3cef2f6e29ee0efda52dd9ea20)