OF-3049: Prevent Session comparison exception
When sorting of ClientSessions takes a while (eg: in a clustered setup with many sessions), the comparator breaks with "java.lang.IllegalArgumentException: Comparison method violates its general contract!" when values of properties used by the Comparator change. This is easily reproduces by comparing sessions based on amounts of data transferred, while those sessions are actively transmitting data.

This commit compares the sessions based on a snapshot of their value of the property used for comparison.

As a result, the comparison may be 'outdated': sorted sessions may appear to be poorly sorted, when the presentation layer uses the up-to-date value, rather than the snapshot. That is preferable to having exceptions.