OF-3055: When removing childelement, don't assume there's only one
There are various bits in the code where an XML snippet (typically an XMPP stanza) is being manipulated by removing a particular child element. This can have various reasons, including:

- removing client-provided data that should be server-provided (to avoid spoofing)
- re-use of an element for other purposes

When an element is removed, a common assumption is that there will be at most one such element in the XML snippet. This is driven mostly by an expectation, not by inspection.

To be safe, remove all occurrences, not just the first one.