OF-3056: Refactor processing of directed presence
When _removing_ a directed presence from memory (which happens when an 'unavailable' is sent), the previous code seems to try to 'optimize' things, by differentating between a presence removal that's registered for a client session, or any other session.

Such an optimization is undesirable, for a number of reasons:
- it requires the client to currently be online;
- it sets a cluster-side lock (through the lookup on RoutingTable)
- it adds quite a bit of code complexity (making maintenance harder)

When _adding_ the directed presence, the distinction is _not_ made, which adds to the believe that there is code smell.

In this commit, the 'optimization' is replaced by a simpler code path.