SessionManager#terminateDetached to use Session rather than Route
The safeguard that is modified here asserts that a to-be-terminated session isn't replaced by another session.

The previous implementation checked if a new _route_ exists for the JID. A route is a more specific session: one that is 'routable' (e.g. has sent initial presence). That distinction is not relevant for the feature that's implemented here.

It is unlikely that this change will result in a functional difference (as having a detached session that's not routable is unlikely), yet change should be a better representation of what is being asserted.