PresenceUpdateHandler#removedExpiredPresences to use Session rather than Route
When determining if data associated to remote entities can be cleaned up, it is better to evaluate if this data is associated to existing _sessions_ than _routes_.

A route is a more specific session: one that is 'routable' (e.g. has sent initial presence). It is conceivable that a directed presence has been sent to a non-routable ('hidden') entity, that shouldn't be removed.