OF-3059: Remove 'anonymous route' concept
Openfire used the definition of an ‘anonymous route’ in various places (notably the routing table and session manager).

An anonymous route should be defined as “a route-able session of which the users has an anonymous authentication token.”

A ‘route’ is not inherently anonymous or non-anonymous, a session is.

Implementation that previously used the 'anonymous route' definition is interested only in the fact if the corresponding session is anonymous, not if the session is route-able.

In this commit, the implementations of methods named `isAnonymousRoute` are deprecated. Their implementation is replaced by checks that evaluate the authentication mechanism used in a session.